align obj/body
